refactor animation in cast credits
