+++ presentation/details/src/main/java/com/giraffe/details/components/PosterHorizontal.kt
@@ -1,4 +1,4 @@
-package com.giraffe.details.components.recomended
+package com.giraffe.details.components
 
 import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.animation.AnimatedVisibilityScope
@@ -72,11 +72,13 @@ fun PosterHorizontal(
                         animatedVisibilityScope = animatedVisibilityScope
                     )
             ) {
-
                 Icon(
                     painter = painterResource(Theme.icons.dueTone.image),
                     contentDescription = stringResource(R.string.loading_image),
-                    modifier = Modifier.size(32.dp),
+                    modifier = Modifier
+                        .fillMaxHeight()
+                        .width(64.dp)
+                        .padding(16.dp),
                     tint = Theme.color.brand.secondary
                 )
 
@@ -106,7 +108,7 @@ fun PosterHorizontal(
                                     animatedVisibilityScope = animatedVisibilityScope
                                 )
                         )
-                        AnimatedVisibility(poster.genres != null) {
+                        AnimatedVisibility(!poster.genres.isNullOrBlank()) {
                             Text(
                                 text = poster.genres.toString(),
                                 style = Theme.textStyle.body.sm.regular,
@@ -126,13 +128,13 @@ fun PosterHorizontal(
                     )
                 }
 
-                AnimatedVisibility(poster.time != null) {
+                AnimatedVisibility(!poster.time.isNullOrBlank()) {
                     IconWithText(
                         icon = painterResource(Theme.icons.dueTone.clock),
                         text = poster.time.toString()
                     )
                 }
-                AnimatedVisibility(poster.date != null) {
+                AnimatedVisibility(!poster.date.isNullOrBlank()) {
                     IconWithText(
                         icon = painterResource(Theme.icons.dueTone.calendar),
                         text = poster.date.toString()
