+++ presentation/details/src/main/java/com/giraffe/details/screens/recommended/series/TransitionBetweenLazyColumnToLazyVerticalGridSeries.kt
@@ -20,8 +20,8 @@ import androidx.compose.foundation.lazy.rememberLazyListState
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.unit.dp
 import androidx.paging.compose.LazyPagingItems
-import com.giraffe.details.components.recomended.PosterHorizontal
-import com.giraffe.details.components.recomended.PosterVertically
+import com.giraffe.details.components.PosterHorizontal
+import com.giraffe.details.components.PosterVertically
 import com.giraffe.details.models.SeriesUi
 import com.giraffe.details.models.toPoster
